refactor: update import paths to client directory
