+++ apps/web/components/client/nav/NavBar.tsx
@@ -4,7 +4,7 @@ import { useEffect, useState } from "react";
 import Link from "next/link";
 import { useRouter } from "next/navigation";
 import CompactLogo from "../logo/CompactLogo";
-import { Button } from "@/components/ui/button/Button";
+import { Button } from "@/components/client/ui/button/Button";
 import type { SessionData } from "@/lib/auth/types";
 
 export default function NavBar() {
